Abort queryAndRun on failure when using parallelLimitFunc
Need to be able to abort queryAndRun when a failure occurs while
remaining backwards compatible with anything currently using
queryAndRun. This commit enables this ability only when the
parallelLimitFunc option is used.

Because of the version of the async library being used, queryAndRun must
be, and can only be aborted when parallelLimitFunc is enabled.