Enable mod_proxy ProxyPass support.
Most changes involve adjusting absolute URLs to include a prepend. The prepend is currently defined as whatever is after the host:port in config.default_url.

This allows DN to run localhost:someport but fixes urls so that they point to somewhere else (going through the proxy), as what may happen if you're using Apache to proxy from an apache instance running other websites to a dreadnot or multiple dreadnots as the case may be.