add a github workflow
This adds a workflow named "test" so that pull requests will be tested
and given a commit status, which is a pre-requisite for merging now.

I also add a shrinkwrap file here because we're currently on node
0.10.33, which doesn't support any other kind of dependency locking
that I know of.  With this file present, an `npm install` will install
the versions specified here instead of in package.json. It effectively
acts as a `package-lock.json` file, but that's introduced much later
in npm's life. Eventually, we'll have to switch over because
shrinkwrap is deprecated in favor of `package-lock.json`. For now,
this will let us ensure we're using the correct depenedency versions
in dev. I pulled this from a current dreadnot production deployment.